+++ src/status.ts
@@ -435,7 +435,7 @@ async function printLocizeFunnel () {
   if (!(await shouldShowFunnel('status'))) return
 
   console.log(chalk.yellow.bold('\n✨ Take your localization to the next level!'))
-  console.log('Manage translations with your team in the cloud with locize => https://www.locize.com/docs/getting-started')
+  console.log('Manage translations with your team in the cloud with Locize => https://www.locize.com/docs/getting-started')
   console.log(`Run ${chalk.cyan('npx i18next-cli locize-migrate')} to get started.`)
 
   return recordFunnelShown('status')